step11-mariadb-practice/requirements.txt
PyMySQL==1.1.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
